requirements.txt
fastapi[all]
uvicorn
sqlmodel
psycopg[binary]
python-dotenv
confluent-kafka